Add support for juvix-standalone code blocks (#80)
This PR sets up a plugin to give support for `juvix` and
`juvix-standalone` code blocks via
[juvix-mkdocs](https://github.com/anoma/juvix-mkdocs). This plugin
effectively addresses the issue of unreliable JavaScript syntax
highlighting for Juvix code blocks that do not depend on another code
block in the same file. To produce the syntax highlighting, we use the
compiler. In particular, we rely on the HTML output generated from a
module created with the content of the given block.

![2023-10-11 00 16
28](https://github.com/anoma/juvix-docs/assets/1428088/c0bdfe79-bee7-4ee6-a4b1-76d6ba0c70ef)