Make all pages accessible without going through the navigation menu (#131)
On many browsers, if the window is too small the navigation menus at the
top and on the left just don't display. Instead, there is just a single
button in the top-left corner which is easy to miss.

Hence, all pages should be accessible without going through the
generated navigation menu. This PR adds more links on the front page and
creates top-level pages with links for "How-to guides", "Reference" and
"Tutorials".